feat: refactor architecture and connect fragment to navigation
