+++ core/navigation/src/main/res/navigation/nav_graph.xml
@@ -1,7 +1,11 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/together"  >
+    app:startDestination="@id/home"  >
+
+    <fragment
+        android:id="@+id/home"
+        android:name="com.capstone.surfingthegangwon.presentation.home.HomeFragment" />
 
     <fragment
         android:id="@+id/together"
