+++ core/ui/src/androidTest/java/com/capstone/surfingthegangwon/core/ui/ExampleInstrumentedTest.kt
@@ -1,10 +1,12 @@
-package com.capstone.surfingthegangwon.presentation.together
+package com.capstone.surfingthegangwon.core.ui
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
 import org.junit.Test
 import org.junit.runner.RunWith
+
+import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -17,9 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals(
-            "com.capstone.surfingthegangwon.presentation.together.test",
-            appContext.packageName
-        )
+        assertEquals("com.capstone.surfingthegangwon.core.ui.test", appContext.packageName)
     }
 }
