+++ core/ui/src/main/AndroidManifest.xml
