+++ core/ui/src/test/java/com/capstone/surfingthegangwon/core/ui/ExampleUnitTest.kt
@@ -1,7 +1,8 @@
-package com.capstone.surfingthegangwon.presentation.main
+package com.capstone.surfingthegangwon.core.ui
 
-import org.junit.Assert.assertEquals
 import org.junit.Test
+
+import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
