+++ presentation/home/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace = "com.capstone.home"
+    namespace = "com.capstone.surfingthegangwon.presentation.home"
     compileSdk = 35
 
     defaultConfig {
@@ -36,7 +36,8 @@ android {
 }
 
 dependencies {
-    implementation(project(":ui"))
+    implementation(project(":core:navigation"))
+    implementation(project(":core:ui"))
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
@@ -47,4 +48,6 @@ dependencies {
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
+    implementation(libs.androidx.navigation.fragment.ktx)
+    implementation(libs.androidx.navigation.ui.ktx)
 }
