+++ presentation/home/src/androidTest/java/com/capstone/surfingthegangwon/presentation/home/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.capstone.surfingthegangwon.presentation.together
+package com.capstone.surfingthegangwon.presentation.home
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,9 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals(
-            "com.capstone.surfingthegangwon.presentation.together.test",
-            appContext.packageName
-        )
+        assertEquals("com.capstone.surfingthegangwon.presentation.home.test", appContext.packageName)
     }
 }
