--- presentation/home/src/main/java/com/capstone/home/HomeViewModel.kt
@@ -1,7 +1,0 @@
-package com.capstone.home
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
