+++ presentation/home/src/main/java/com/capstone/surfingthegangwon/presentation/home/HomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.capstone.surfingthegangwon.presentation.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
