+++ presentation/home/src/main/res/layout/fragment_home.xml
@@ -4,9 +4,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".HomeFragment">
+    tools:context="com.capstone.surfingthegangwon.presentation.home.HomeFragment">
 
-    <com.capstone.ui.CustomHeaderView
+    <com.capstone.surfingthegangwon.core.ui.CustomHeaderView
         android:id="@+id/headerView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
