+++ presentation/together/src/main/res/layout/fragment_together.xml
@@ -6,12 +6,22 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".TogetherFragment">
 
-    <com.capstone.ui.CustomHeaderView
+    <com.capstone.surfingthegangwon.core.ui.CustomHeaderView
         android:id="@+id/headerView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"/>
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="TextView"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
