+++ presentation/together/src/test/java/com/capstone/surfingthegangwon/presentation/together/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.capstone.surfingthegangwon.presentation.together
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
