+++ settings.gradle.kts
@@ -27,5 +27,5 @@ include(":data:together")
 include(":domain:together")
 include(":core:resource")
 include(":core:navigation")
-include(":ui")
 include(":presentation:home")
+include(":core:ui")
